Add the ability to run the specs on travis
The before_install.sh script installs the proper version of postgres
and then installs pglogical based on the distro version we get.

We also need to explicitly specify the spec script because we need
to run the setup to create the databases before we run the spec itself.

We force the distro to trusty in .travis.yml because pglogical doesn't
ship packages for earlier versions