Account for symbol table names in drop_table patch
The array of tables names coming from the database contains only
strings.

When comparing the given name to the list we were getting a false
negative result and not removing the table from the replication
set before trying to drop it.

Fixes https://bugzilla.redhat.com/show_bug.cgi?id=1540528